rename newappimage
